+++ Main/Source/human.cpp
@@ -1626,7 +1626,7 @@ void humanoid::DrawSilhouette(truth AnimationDraw) const
 		  { 0, 0 },
 		  { 0, 0 },
 		  { TILE_SIZE, TILE_SIZE },
-		  { ivanconfig::GetContrastLuminance() },
+		  { static_cast<int>(ivanconfig::GetContrastLuminance()) },
 		  TRANSPARENT_COLOR,
 		  ALLOW_ANIMATE };
 
@@ -5281,7 +5281,7 @@ void petrusswife::BeTalkedTo()
   int Accepted = 0;
   truth RefusedSomething = false;
 
-  for(int c = 0; c < Item.size(); ++c)
+  for(size_t c = 0; c < Item.size(); ++c)
     if(!MakesBurdened(GetCarriedWeight() + Item[c]->GetWeight()))
     {
       ++Accepted;
@@ -5310,7 +5310,7 @@ void guard::BeTalkedTo()
      || Item.empty())
     humanoid::BeTalkedTo();
 
-  for(int c = 0; c < Item.size(); ++c)
+  for(size_t c = 0; c < Item.size(); ++c)
   {
     Item[c]->RemoveFromSlot();
     GetStack()->AddItem(Item[c]);
